Rework notification and mailbox loading
