+++ client/my-sites/email/controller.js
@@ -212,7 +212,6 @@ export default {
 					selectedDomainName={ pageContext.params.domain }
 					selectedEmailProviderSlug={ pageContext.query.provider }
 					selectedIntervalLength={ castIntervalLength( pageContext.query.interval ) }
-					enableSetupRedirects
 				/>
 			</CalypsoShoppingCartProvider>
 		);
