+++ client/my-sites/email/email-management/email-home.tsx
@@ -71,7 +71,6 @@ interface EmailManagementHomeProps {
 	selectedIntervalLength?: IntervalLength;
 	showActiveDomainList?: boolean;
 	source: string;
-	enableSetupRedirects: boolean;
 }
 
 const domainHasEmail = ( domain: ResponseDomain ) =>
@@ -86,7 +85,6 @@ const EmailHome = ( props: EmailManagementHomeProps ) => {
 		selectedIntervalLength,
 		sectionHeaderLabel,
 		source,
-		enableSetupRedirects = true,
 	} = props;
 
 	const selectedSite = useSelector( getSelectedSite );
@@ -179,7 +177,7 @@ const EmailHome = ( props: EmailManagementHomeProps ) => {
 		);
 	}
 
-	if ( isSingleDomainThatHasEmail && enableSetupRedirects ) {
+	if ( isSingleDomainThatHasEmail ) {
 		if (
 			( domainsWithEmail[ 0 ].titanMailSubscription?.maximumMailboxCount ?? 0 ) > 0 &&
 			getConfiguredTitanMailboxCount( domainsWithEmail[ 0 ] ) === 0
